Extract non-PySNMP specific classes and types
This extracts from `zino.trapd.pysnmp_backend` all auxiliary classes and
type definitions that are not specific to the PySNMP back-end and moves
them to `zino.trapd.base` so they can be re-used by other back-ends.
