+++ src/zino/trapobservers/ignored_traps.py
@@ -3,7 +3,7 @@
 import logging
 from typing import Optional, Set
 
-from zino.trapd import TrapMessage, TrapObserver, TrapType
+from zino.trapd.base import TrapMessage, TrapObserver, TrapType
 
 _logger = logging.getLogger(__name__)
